vision/requirements.txt
fastapi
uvicorn
python-dotenv

opencv-python
ultralytics>=8.1.0
supervision
ftfy
regex
python-multipart

pyrealsense2

--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision